Refactored `rusty-junctions-partial-pattern-derive`
Refactored `rusty-junctions-partial-pattern-derive` into the
`rusty-junctions-macro` crate as a module instead of a separate crate.
