--- rusty-junctions-macro/rusty-junctions-partial-pattern-derive/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "rusty-junctions-partial-pattern-derive"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-syn = { version = "1.0", features = ["full"] }
-quote = "1.0"
-proc-macro2 = "1.0"
-rusty-junctions-utils = { path = "./../rusty-junctions-utils" }
-
-[lib]
-proc-macro = true
